+++ components/svg/assets.tsx
@@ -5,14 +5,12 @@ import { SVGProps } from './svg.types';
 const Assets: FC<SVGProps> = ({ maxWidth, maxHeight, ...props }) => (
   <svg
     style={{ maxWidth, maxHeight }}
-    viewBox="0 0 32 32"
+    viewBox="0 0 11 10"
     fill="none"
     {...props}
   >
     <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M16 4.5C13.7909 4.5 12 6.29086 12 8.5C12 10.7091 13.7909 12.5 16 12.5C18.2091 12.5 20 10.7091 20 8.5C20 6.29086 18.2091 4.5 16 4.5ZM10 8.5C10 5.18629 12.6863 2.5 16 2.5C19.3137 2.5 22 5.18629 22 8.5C22 11.8137 19.3137 14.5 16 14.5C12.6863 14.5 10 11.8137 10 8.5ZM23.5 17.5C21.2909 17.5 19.5 19.2909 19.5 21.5C19.5 23.7091 21.2909 25.5 23.5 25.5C25.7091 25.5 27.5 23.7091 27.5 21.5C27.5 19.2909 25.7091 17.5 23.5 17.5ZM17.5 21.5C17.5 18.1863 20.1863 15.5 23.5 15.5C26.8137 15.5 29.5 18.1863 29.5 21.5C29.5 24.8137 26.8137 27.5 23.5 27.5C20.1863 27.5 17.5 24.8137 17.5 21.5ZM4.5 21.5C4.5 19.2909 6.29086 17.5 8.5 17.5C10.7091 17.5 12.5 19.2909 12.5 21.5C12.5 23.7091 10.7091 25.5 8.5 25.5C6.29086 25.5 4.5 23.7091 4.5 21.5ZM8.5 15.5C5.18629 15.5 2.5 18.1863 2.5 21.5C2.5 24.8137 5.18629 27.5 8.5 27.5C11.8137 27.5 14.5 24.8137 14.5 21.5C14.5 18.1863 11.8137 15.5 8.5 15.5Z"
+      d="M5.24998 0.40625C5.54119 0.40625 5.80984 0.560059 5.9575 0.812305L10.3872 8.35918C10.5369 8.61348 10.5369 8.92725 10.3913 9.18154C10.2457 9.43584 9.97293 9.59375 9.67966 9.59375H0.820288C0.527026 9.59375 0.254272 9.43584 0.108667 9.18154C-0.0369386 8.92725 -0.0348878 8.61143 0.112768 8.35918L4.54246 0.812305C4.69011 0.560059 4.95876 0.40625 5.24998 0.40625ZM5.24998 3.03125C4.97722 3.03125 4.75779 3.25068 4.75779 3.52344V5.82031C4.75779 6.09307 4.97722 6.3125 5.24998 6.3125C5.52273 6.3125 5.74216 6.09307 5.74216 5.82031V3.52344C5.74216 3.25068 5.52273 3.03125 5.24998 3.03125ZM5.90623 7.625C5.90623 7.45095 5.83709 7.28403 5.71401 7.16096C5.59094 7.03789 5.42402 6.96875 5.24998 6.96875C5.07593 6.96875 4.90901 7.03789 4.78594 7.16096C4.66287 7.28403 4.59373 7.45095 4.59373 7.625C4.59373 7.79905 4.66287 7.96597 4.78594 8.08904C4.90901 8.21211 5.07593 8.28125 5.24998 8.28125C5.42402 8.28125 5.59094 8.21211 5.71401 8.08904C5.83709 7.96597 5.90623 7.79905 5.90623 7.625Z"
       fill="currentColor"
     />
   </svg>
