+++ pages/_app.tsx
@@ -1,8 +1,11 @@
+import 'react-loading-skeleton/dist/skeleton.css';
+
 import { Global } from '@emotion/react';
 import { NextPage } from 'next';
 import type { AppProps } from 'next/app';
 import { ReactNode, StrictMode } from 'react';
 import { Toaster } from 'react-hot-toast';
+import { SkeletonTheme } from 'react-loading-skeleton';
 
 import ModalProvider from '@/components/modal-provider';
 import { TOAST_DURATION } from '@/constants';
@@ -13,7 +16,6 @@ const MyApp = ({ Component, pageProps }: AppProps<NextPage>): ReactNode => (
     <StrictMode>
       <Global styles={GlobalStyles} />
       <ModalProvider />
-
       <Toaster
         position="bottom-right"
         toastOptions={{
@@ -29,7 +31,9 @@ const MyApp = ({ Component, pageProps }: AppProps<NextPage>): ReactNode => (
           },
         }}
       />
-      <Component {...pageProps} />
+      <SkeletonTheme baseColor="#99BBFF28" highlightColor="#99BBFF14">
+        <Component {...pageProps} />
+      </SkeletonTheme>
     </StrictMode>
   </>
 );
