+++ views/pool-create/components/pair-token/input/components/balance.tsx
