+++ views/pool-create/components/pair-token/input/components/dollar-value.tsx
@@ -6,7 +6,7 @@ import { useFormContext, useWatch } from 'react-hook-form';
 import { useCoinsPrice } from '@/hooks/use-coins-price';
 import { formatDollars } from '@/utils';
 
-import { CreatePoolForm } from '../../../pool-create.types';
+import { CreatePoolForm } from '../../../../pool-create.types';
 import { InputProps } from '../input.types';
 
 const AmountInDollar: FC<InputProps> = ({ index }) => {
