+++ views/pool-create/components/pair-token/input/components/header-info.tsx
