+++ views/pool-create/components/pair-token/input/input.types.ts
