+++ views/pool-create/components/pool-error-section/error-section.types.ts
@@ -1,0 +1,7 @@
+export enum CreatePoolErrorMessagesEnum {
+  sameToken = 'Same token selected',
+  insufficientBalance = 'Insufficient balance',
+  leastOneMove = 'Need to keep at least 1 MOVE',
+  invalidAmount = 'Initial price must be greater than 0',
+  notEnoughToken = 'Not enough token amount',
+}
