+++ views/pool-create/components/pool-error-section/index.tsx
@@ -1,0 +1,15 @@
+import { FC } from 'react';
+
+import { CreatePoolErrorManager } from './pool-error-manager';
+import CreatePoolErrorMessage from './pool-error-message';
+
+const CreatePoolError: FC = () => {
+  return (
+    <>
+      <CreatePoolErrorManager />
+      <CreatePoolErrorMessage />
+    </>
+  );
+};
+
+export default CreatePoolError;
